datapunk/lib/mesh/circuit_breaker/circuit_breaker_manager.py
"""
Service Mesh Circuit Breaker Manager

Coordinates circuit breaker behavior across the Datapunk service mesh,
providing centralized management and monitoring of service reliability
patterns. Enables consistent failure handling across the mesh.

Key features:
- Centralized circuit breaker management
- Dynamic configuration updates
- Cross-service failure correlation
- Real-time state monitoring
- Failure pattern analysis

See sys-arch.mmd Reliability/CircuitBreakerManager for implementation details.
"""

from typing import Dict, Optional, TYPE_CHECKING
import structlog
from datetime import datetime

from datapunk.lib.exceptions import CircuitBreakerError
from .circuit_breaker_strategies import CircuitBreakerStrategy
from .circuit_breaker import CircuitBreaker

if TYPE_CHECKING:
    from datapunk_shared.monitoring import MetricsClient
    from datapunk_shared.cache import CacheClient

logger = structlog.get_logger()

class CircuitBreakerManager:
    """
    Centralized circuit breaker coordination system.
    
    Manages circuit breaker instances across the service mesh,
    providing consistent failure handling and pattern detection.
    Designed for dynamic service environments with varying
    reliability requirements.
    
    TODO: Add support for cross-service failure correlation
    TODO: Implement adaptive threshold adjustment
    
    NOTE: Circuit breaker instances are created lazily to
    optimize resource usage in large service meshes.
    """

logger = structlog.get_logger() 